refactor: introduce counting semaphore (#653)
Add a counting semaphore for the startup for the engines, up to 16 engines can be spawned simultaneously before having to wait for other engines to finish their initialization.initialization. This allows us to set a lower timeout threshold for the startup time on high core count machines, which will lag behind when spawning hundreds of engines near instantaneously.