feat: A few enhancements to the TB adjudication feature (#819)
 - Add support for cutechess options `-tbpieces` and `-tbignore50`.
 - Tweak adjudication wording to be identical to cutechess.
 - Add TB adjudication to the cutechess compatibility test, as suggested by @robertnurnberg on #818.

The modified cutechess compatibility test will look for 345-men TBs and use them if present. Though because of the `-tbpieces 4` option only the 34-men TBs are actually used for adjudication. The modified workflow will download the 345-men TBs for the test to use, and will cache the files if possible.

The `-tbignore50` option can't be tested via the compatibility test because cursed wins/losses are too rare an occurrence. A new unit test is added instead, and a single 5-men WDL file was added to support this (~3.4 MB).